requirements.txt
aiohttp==3.6.2
altgraph==0.17
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
auto-py-to-exe==2.7.5
backcall==0.2.0
beautifulsoup4==4.9.0
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colors.py==0.2.2
cryptography==3.1.1
cx-Freeze==6.2
cycler==0.10.0
decorator==4.4.2
discord==1.0.1
discord.py==1.7.1
distlib==0.3.1
docopt==0.6.2
ebaysdk==2.2.0
Eel==0.12.4
env==0.1.0
filelock==3.0.12
Flask==1.1.2
freeze==3.0
future==0.18.2
gcloud==0.17.0
gevent==20.6.2
gevent-websocket==0.10.1
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-images-download==2.8.0
google-search==1.0.2
googleapis-common-protos==1.52.0
greenlet==0.4.16
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.7.0
jws==0.1.3
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
marshmallow==3.8.0
multidict==4.7.5
oauth2client==3.0.0
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
pipreqs==0.4.10
platformio==5.0.1
praw==7.1.0
prawcore==1.4.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycolors==0.1.2
pycparser==2.20
pycryptodome==3.4.3
pyelftools==0.26
pyfiglet==0.8.post1
Pygments==2.7.2
PyInstaller==3.6
PyJWT==1.7.1
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.3.1
Pyrebase==3.0.27
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-jwt==2.0.1
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
pyzmq==20.0.0
requests==2.24.0
requests-toolbelt==0.7.0
rsa==4.6
semantic-version==2.8.5
six==1.15.0
soupsieve==2.0
tabulate==0.8.7
termcolor==1.1.0
tornado==6.1
traitlets==5.0.5
update-checker==0.17
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
whichcraft==0.6.1
yarg==0.1.9
yarl==1.4.2
youtube-dl==2021.4.7
zope.event==4.4
zope.interface==5.1.0
